+++ src/test/java/com/poorlex/poorlex/expenditure/domain/ExpenditureRepositoryTest.java
@@ -19,11 +19,11 @@ class ExpenditureRepositoryTest extends UsingDataJpaTest implements ReplaceUnder
         final Long memberId = 1L;
         final LocalDate date = LocalDate.now();
 
-        expenditureRepository.save(ExpenditureFixture.simpleWithMainImage(1000, memberId, date));
-        expenditureRepository.save(ExpenditureFixture.simpleWithMainImage(2000, memberId, date));
+        expenditureRepository.save(ExpenditureFixture.simpleWithMainImage(1000L, memberId, date));
+        expenditureRepository.save(ExpenditureFixture.simpleWithMainImage(2000L, memberId, date));
 
         //when
-        final int sumExpenditure = expenditureRepository.findSumExpenditureByMemberIdAndBetween(
+        final Long sumExpenditure = expenditureRepository.findSumExpenditureByMemberIdAndBetween(
                 memberId,
                 date.minusDays(1),
                 date
